LoraWan: rename header files to *.h
