+++ detector/lib/LoraWan/LoraWanAbp.cpp
@@ -1,5 +1,5 @@
-#include "LoraWanAbp.hpp"
 #include <Arduino.h>
+#include <LoraWanAbp.h>
 #undef max
 #undef min
 #include <lmic.h>
