+++ detector/lib/LoraWan/LoraWanAbp.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 #include <ILoraWanConfigAdapter.h>
 
-#include <LoRaWanDriver.hpp>
+#include <LoRaWanDriver.h>
 
 #ifdef __cplusplus
    extern "C"
@@ -29,7 +29,7 @@ void loop_once();
  }
 #endif
 
-#include <LoRaWanDriver.hpp>
+#include <LoRaWanDriver.h>
 
 class LoraWanAbp : public LoRaWanDriver
 {
