+++ detector/platformio.ini
@@ -44,7 +44,7 @@ framework = arduino
 lib_deps =
   ${common.lib_deps_builtin}
   ${common.lib_deps_external}
-upload_port = COM9
+upload_port = COM6
 
 [env:due]
 platform = atmelsam
